PascalsTriangle python
I made this almost 2 years ago.
This is what I did sometimes during the pandemic.
Just woke up and though "how long would it take to crank out a pascal triangle in python?"
If I remember correctly it was 10-20 minutes and I got to bored to format the output in a pretty way. I might do that later. Probably not.
Included is an example of the output (output.pascal)